pyproject.toml
[tool.poetry]
name = "youtube-title-generator"
version = "0.1.0"
description = ""
authors = ["sergio.sampayo"]

[tool.poetry.dependencies]
python = "^3.10"
kaggle = "^1.6.17"
ipython = "^8.27.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
elasticsearch = "^8.15.1"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
openai = "^1.47.0"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
poetry-plugin-dotenv = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"